chore: move out the initialization helpers to a module
