+++ packages/notifications/src/inAppMessaging/providers/pinpoint/utils/helpers.ts
@@ -21,11 +21,6 @@ import { resolveConfig } from './resolveConfig';
 import { resolveCredentials } from './resolveCredentials';
 import { CATEGORY } from './constants';
 import { getInAppMessagingUserAgentString } from './userAgent';
-import {
-	InAppMessagingValidationErrorCode,
-	assertValidationError,
-} from '../../../errors';
-import { initialized } from '../apis/initializeInAppMessaging';
 
 const DELIVERY_TYPE = 'IN_APP_MESSAGE';
 
@@ -35,12 +30,6 @@ let eventMetricsMemo = {};
 
 export const logger = new ConsoleLogger('InAppMessaging.Pinpoint.Utils');
 
-export function assertInitializationError() {
-	assertValidationError(
-		initialized,
-		InAppMessagingValidationErrorCode.NotInitialized
-	);
-}
 export const recordAnalyticsEvent = (
 	event: PinpointMessageEvent,
 	message: InAppMessage
