+++ packages/notifications/src/inAppMessaging/utils/index.ts
@@ -1,0 +1,8 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+export {
+	isInAppMessagingInitialized,
+	assertInitializationError,
+	setInAppMessagingInitializedStatus,
+} from './statusHelpers';
